Add the middleware
