+++ src/http/conversions.rs
@@ -1,32 +1,5 @@
 use crate::models::api::external::token::ExternalToken;
 use actix_web::http::header::HeaderValue;
-use anyhow::bail;
-
-impl TryFrom<&HeaderValue> for ExternalToken {
-    type Error = anyhow::Error;
-
-    // Compiler will complain that `return Err` statements in this code are unreachable,
-    // but this is not true. See test cases below.
-    #[allow(unreachable_code)]
-    fn try_from(value: &HeaderValue) -> Result<Self, Self::Error> {
-        match value.to_str() {
-            Ok(string_value) => {
-                let tokens = string_value.split(" ").collect::<Vec<&str>>();
-
-                if tokens.len() != 2 {
-                    return Err(bail!("Invalid token format"));
-                }
-
-                if tokens[0] != "Bearer" {
-                    return Err(bail!("Invalid token format"));
-                }
-
-                Ok(ExternalToken::from(tokens[1].to_owned()))
-            }
-            Err(_) => bail!("Invalid token format"),
-        }
-    }
-}
 
 #[cfg(test)]
 mod tests {
