+++ src/http/middleware.rs
@@ -1,1 +1,2 @@
+mod external_token;
 pub mod extract_external_token;
